+++ lib/ui/page/erase/view.dart
@@ -125,6 +125,8 @@ class EraseView extends StatelessWidget {
             label: 'label_identifier'.l10n,
             hint: 'label_sign_in_input'.l10n,
             floatingLabelBehavior: FloatingLabelBehavior.always,
+            spellCheck: false,
+            autocomplete: AutocompleteKind.username,
           ),
           const SizedBox(height: 16),
           ReactiveTextField.password(