+++ lib/ui/widget/text_field.dart
@@ -72,6 +72,8 @@ class ReactiveTextField extends StatelessWidget {
     this.selectable,
     this.floatingAccent = false,
     this.textCapitalization = TextCapitalization.sentences,
+    this.spellCheck = true,
+    this.autocomplete,
   });
 
   /// Constructs a [ReactiveTextField] tuned best for password-type fields.
@@ -83,6 +85,7 @@ class ReactiveTextField extends StatelessWidget {
     required RxBool obscured,
     bool treatErrorAsStatus = true,
     TextStyle? style,
+    AutocompleteKind autocomplete = AutocompleteKind.currentPassword,
   }) {
     return Obx(() {
       return ReactiveTextField(
@@ -101,6 +104,8 @@ class ReactiveTextField extends StatelessWidget {
         treatErrorAsStatus: treatErrorAsStatus,
         textCapitalization: TextCapitalization.none,
         style: style,
+        spellCheck: false,
+        autocomplete: autocomplete,
       );
     });
   }
@@ -243,6 +248,13 @@ class ReactiveTextField extends StatelessWidget {
 
   /// Indicator whether the style of floating [label] should have accent.
   final bool floatingAccent;
+
+  /// Indicator whether spell checking and auto correction should be enabled for
+  /// this field.
+  final bool spellCheck;
+
+  /// [AutocompleteKind] to pass to this field.
+  final AutocompleteKind? autocomplete;
 
   @override
   Widget build(BuildContext context) {
@@ -462,6 +474,12 @@ class ReactiveTextField extends StatelessWidget {
               textInputAction: textInputAction,
               textCapitalization: textCapitalization,
               maxLength: maxLength,
+              spellCheckConfiguration: spellCheck
+                  ? null
+                  : SpellCheckConfiguration.disabled(),
+              autocorrect: spellCheck ? null : false,
+              enableSuggestions: spellCheck,
+              autofillHints: [?autocomplete?.value],
               contextMenuBuilder: (_, field) {
                 final double dx = field.contextMenuAnchors.primaryAnchor.dx;
                 final double dy = field.contextMenuAnchors.primaryAnchor.dy;
@@ -550,6 +568,22 @@ class ReactiveTextField extends StatelessWidget {
         ),
       );
     });
+  }
+}
+
+/// Possible autocomplete hints to pass to a [TextField].
+enum AutocompleteKind {
+  currentPassword,
+  newPassword,
+  username;
+
+  /// Returns the actual value to use from this [AutocompleteKind].
+  String get value {
+    return switch (this) {
+      currentPassword => 'current-password',
+      newPassword => 'new-password',
+      username => 'username',
+    };
   }
 }