+++ lib/ui/page/home/widget/direct_link.dart
@@ -1,4 +1,4 @@
-// Copyright © 2022-2025 IT ENGINEERING MANAGEMENT INC,
+// Copyright © 2022-2026 IT ENGINEERING MANAGEMENT INC,
 //                       <https://github.com/team113>
 //
 // This program is free software: you can redistribute it and/or modify it under
@@ -126,6 +126,7 @@ class _DirectLinkFieldState extends State<DirectLinkField> {
             floatingAccent: true,
             label: Config.link,
             floatingLabelBehavior: FloatingLabelBehavior.always,
+            spellCheck: false,
           ),
           SizedBox(height: 8),
           Stack(