+++ lib/util/media_utils.dart
@@ -1,4 +1,4 @@
-// Copyright © 2022-2025 IT ENGINEERING MANAGEMENT INC,
+// Copyright © 2022-2026 IT ENGINEERING MANAGEMENT INC,
 //                       <https://github.com/team113>
 //
 // This program is free software: you can redistribute it and/or modify it under
@@ -351,7 +351,14 @@ class MediaUtilsImpl {
           DisplayVideoTrackConstraints();
       if (screen.device != null) constraints.deviceId(screen.device!);
       constraints.idealFrameRate(screen.framerate ?? 30);
+      constraints.idealHeight(1080);
       settings.displayVideo(constraints);
+
+      if (screen.audio) {
+        final DisplayAudioTrackConstraints constraints =
+            DisplayAudioTrackConstraints();
+        settings.displayAudio(constraints);
+      }
     }
 
     return settings;
@@ -403,10 +410,13 @@ class VideoPreferences extends TrackPreferences {
 
 /// [TrackPreferences] of a screen share track.
 class ScreenPreferences extends TrackPreferences {
-  const ScreenPreferences({super.device, this.framerate});
+  const ScreenPreferences({super.device, this.framerate, this.audio = false});
 
   /// Preferred framerate of the screen track.
   final int? framerate;
+
+  /// Indicator whether audio system capture should be enabled.
+  final bool audio;
 }
 
 /// Extension adding conversion on [MediaDeviceDetails] to [AudioSpeakerKind].