+++ lib/domain/service/my_user.dart
@@ -129,6 +129,10 @@ class MyUserService extends DisposableService {
       }
 
       await WebUtils.protect(() async {
+        if (isClosed) {
+          return;
+        }
+
         await _myUserRepository.updateUserPassword(oldPassword, newPassword);
 
         // TODO: Replace `unsafe` with something more granular and correct.