+++ lib/store/call.dart
@@ -977,6 +977,10 @@ class CallRepository extends DisposableInterface
     _remoteSubscription?.cancel(immediate: true);
 
     await WebUtils.protect(() async {
+      if (isClosed) {
+        return;
+      }
+
       _remoteSubscription = StreamQueue(_incomingEvents(count));
       await _remoteSubscription!.execute(_incomingChatCallsTopEvent);
     }, tag: 'incomingCalls');